backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.12.1
httpx==0.25.2
email-validator==2.1.0
numpy==1.26.2
requests==2.31.0
python-multipart==0.0.6
